+++ internal/provider/schema_cloudflare_custom_hostname.go
@@ -134,6 +134,14 @@ func resourceCloudflareCustomHostnameSchema() map[string]*schema.Schema {
 				},
 			},
 		},
+		"custom_metadata": {
+			Type:        schema.TypeMap,
+			Optional:    true,
+			Description: "Custom metadata associated with custom hostname. Only supports primitive string values, all other values are accessible via the API directly.",
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
 		"status": {
 			Type:        schema.TypeString,
 			Computed:    true,